Revert "sdm660-common: Implement FastCharge HAL"
Iafa7ce24b9f82f44f03dd3ff04bef72dda59cdb6 brings back prebuilt Health HAL,
so this is not needed anymore. Also the fastcharge implementation is kinda
janky on this platform, so this will be dropped until we come up with something
better. A toggle is no longer needed just for fast charging anymore, as this will
now be handled by the Health HAL.

This reverts commit 1d06c0590b35ed2ac8be5248193a48de0ac54b42.

Change-Id: Ifae7aa9e4b3e815abad90ffb034bf1e114605aa7